Ahh, I get it now.
They were talking about the stack contents on entering our function, not an
expectation of the kernel. It is consistent.

And the dietlibc comment about stacks was just saying we can't pop anything
off the stack and expect its value to be consistent later, because when a
signal arrives the new stack frame would clobber anything just off the stack.