Single-threaded/multi-threaded split. Only tested so far under Darwin/ppc (and lightly, at that).
I'm optimistic that a single-threaded FreeBSD version will work.

Also got rid of the separate builds for debug/optimized. It didn't make a
difference for the performance-sensitive code, so I'll just stick with debug.
There's a flag still for the debug defines (that write [S] and [J] on signal
receipt). But I expect that once a platform is working, it will just be
toggled off. No need for separate builds.